:lipstick: warning fixes

+++ features/minimizer/MultiDimFile.cxx
@@ -38,7 +38,7 @@ namespace Hal {
       fTree->GetEntry(0);
       fMultiOffsets = fDataManager->GetMultiFactors();
 
-      if (fValues->size() != fDataManager->GetParametersNo() + 1) {
+      if ((int) fValues->size() != fDataManager->GetParametersNo() + 1) {
         Hal::Cout::PrintInfo(Form("Ooops, incompatible sizes in MultiDimFile expected %i found %i",
                                   fDataManager->GetParametersNo(),
                                   (int) fValues->size()),
